♻️ (NewGEK): Refactor DnDAnswerNode
